SYS: Fixed linter errors and resulting input parameter changes in respective header file.

+++ Libraries/PeriphDrivers/Include/MAX32665/mxc_delay.h
@@ -57,7 +57,7 @@
  * x = SEC(3) // 3 seconds -> x = 3,000,000 
  * \endcode
  */
-#define MXC_DELAY_SEC(s) (((unsigned long)s) * 1000000UL)
+#define MXC_DELAY_SEC(s) (((uint32_t)s) * 1000000UL)
 /**
  * Macro used to specify a microsecond timing parameter in milliseconds.
  * \code
@@ -91,7 +91,7 @@ typedef void (*mxc_delay_complete_t)(int result);
  * @param      us    microseconds to delay
  * @return     #E_NO_ERROR if no errors, @ref MXC_Error_Codes "error" if unsuccessful.
  */
-int MXC_Delay(unsigned long us);
+int MXC_Delay(uint32_t us);
 
 /**
  * @brief      Starts a non-blocking delay for the specified number of
@@ -106,7 +106,7 @@ int MXC_Delay(unsigned long us);
  * @return     #E_NO_ERROR if no errors, #E_BUSY if currently servicing another
  *             delay request.
  */
-int MXC_DelayAsync(unsigned long us, mxc_delay_complete_t callback);
+int MXC_DelayAsync(uint32_t us, mxc_delay_complete_t callback);
 
 /**
  * @brief      Returns the status of a non-blocking delay request
