+++ Libraries/PeriphDrivers/Include/MAX78002/mxc_delay.h
@@ -57,7 +57,7 @@
  * x = SEC(3) // 3 seconds -> x = 3,000,000
  * \endcode
  */
-#define MXC_DELAY_SEC(s) (((unsigned long)s) * 1000000UL)
+#define MXC_DELAY_SEC(s) (((uint32_t)s) * 1000000UL)
 /**
  * Macro used to specify a microsecond timing parameter in milliseconds.
  * \code
@@ -90,10 +90,10 @@ typedef void (*mxc_delay_complete_t)(int result);
  * @param      us    microseconds to delay
  * @return     #E_NO_ERROR if no errors, @ref MXC_Error_Codes "error" if unsuccessful.
  */
-int MXC_Delay(unsigned long us);
+int MXC_Delay(uint32_t us);
 
 // AI87-TODO: Merged from Github, need to add comments.
-int MXC_DelayAsync(unsigned long us, mxc_delay_complete_t callback);
+int MXC_DelayAsync(uint32_t us, mxc_delay_complete_t callback);
 int MXC_DelayCheck(void);
 void MXC_DelayAbort(void);
 
@@ -109,7 +109,7 @@ void MXC_DelayAbort(void);
  * @param      us    microseconds to delay
  * @return     #E_NO_ERROR if no errors, @ref MXC_Error_Codes "error" if unsuccessful.
  */
-int MXC_Delay(unsigned long us);
+int MXC_Delay(uint32_t us);
 
 /**
  * @brief      Starts a non-blocking delay for the specified number of
@@ -124,7 +124,7 @@ int MXC_Delay(unsigned long us);
  * @return     #E_NO_ERROR if no errors, #E_BUSY if currently servicing another
  *             delay request.
  */
-int MXC_DelayAsync(unsigned long us, mxc_delay_complete_t callback);
+int MXC_DelayAsync(uint32_t us, mxc_delay_complete_t callback);
 
 /**
  * @brief      Returns the status of a non-blocking delay request
